+++ domain/entities/Sample.go
@@ -2,7 +2,7 @@ package entities
 
 // Sample entity
 type Sample struct {
-	ID       int64
+	ID       int
 	Name     string
 	Password string
 }
